refactor: AuthResponse 패키지 분리
