+++ src/main/java/nextstep/member/infra/dto/AuthResponse.java
@@ -1,0 +1,14 @@
+package nextstep.member.infra.dto;
+
+public class AuthResponse {
+
+    private String email;
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+}
